+++ backend/mockingbird/src/main/scala/ru/tinkoff/tcb/mockingbird/model/GrpcMethodDescription.scala
@@ -7,6 +7,8 @@ import derevo.circe.encoder
 import derevo.derive
 import eu.timepit.refined.types.string.NonEmptyString
 import io.circe.refined.*
+import io.estatico.newtype.macros.newtype
+import io.estatico.newtype.ops.*
 import io.scalaland.chimney.dsl.TransformationOps
 import sttp.tapir.codec.refined.*
 import sttp.tapir.derevo.schema
@@ -35,20 +37,21 @@ final case class GrpcMethodDescription(
 )
 
 object GrpcMethodDescription {
-  def getRootFields(className: String, definition: GrpcProtoDefinition): IO[ValidationError, List[GrpcField]] = {
-    val name = definition.`package`.map(p => className.stripPrefix(p ++ ".")).getOrElse(className)
+  def getRootFields(
+      name: NormalizedTypeName,
+      types: Map[NormalizedTypeName, GrpcRootMessage] = Map.empty
+  ): IO[ValidationError, List[GrpcField]] =
     for {
-      rootMessage <- ZIO.getOrFailWith(ValidationError(Vector(s"Root message '$className' not found")))(
-        definition.schemas.find(_.name == name)
+      rootMessage <- ZIO.getOrFailWith(ValidationError(Vector(s"Root message '${name}' not found")))(
+        types.get(name)
       )
       rootFields <- rootMessage match {
         case GrpcMessageSchema(_, fields, oneofs, _, _) =>
           ZIO.succeed(fields ++ oneofs.map(_.flatMap(_.options)).getOrElse(List.empty))
         case GrpcEnumSchema(_, _) =>
-          ZIO.fail(ValidationError(Vector(s"Enum cannot be a root message, but '$className' is")))
+          ZIO.fail(ValidationError(Vector(s"Enum cannot be a root message, but '${name}' is")))
       }
     } yield rootFields
-  }
 
   def fromCreateRequest(
       request: CreateGrpcStubRequest,
@@ -56,7 +59,7 @@ object GrpcMethodDescription {
       responseSchema: GrpcProtoDefinition,
       created: Instant,
   ): GrpcMethodDescription = {
-    val proxyUrl = (GProxyResponse.prism >> GProxyResponse.endpoint).getOption(request.response)
+    val proxyUrl = (GProxyResponse.prism >> GProxyResponse.endpoint).getOption(request.response).flatten
 
     request
       .into[GrpcMethodDescription]
@@ -69,6 +72,38 @@ object GrpcMethodDescription {
       .withFieldConst(_.created, created)
       .transform
   }
+
+  @newtype class PackagePrefix private (val asString: String) {
+    def :+(nested: String): PackagePrefix =
+      (asString ++ nested.dropWhile(_ == '.') ++ ".").coerce
+
+    def resolve(n: String): NormalizedTypeName =
+      if (n.startsWith(".")) NormalizedTypeName(n)
+      else if (n.startsWith(asString.drop(1))) NormalizedTypeName(s".$n")
+      else NormalizedTypeName(s"$asString$n")
+  }
+
+  object PackagePrefix {
+    def apply(definition: GrpcProtoDefinition): PackagePrefix =
+      definition.`package`.map(p => s".$p.").getOrElse(".").coerce
+  }
+
+  @newtype class NormalizedTypeName private (val asString: String)
+
+  object NormalizedTypeName {
+    def apply(name: String): NormalizedTypeName =
+      s".${name.dropWhile(_ == '.')}".coerce
+  }
+
+  def makeDictTypes(p: PackagePrefix, ms: Seq[GrpcRootMessage]): Vector[(NormalizedTypeName, GrpcRootMessage)] =
+    ms.foldLeft(Vector.empty[(NormalizedTypeName, GrpcRootMessage)]) {
+      case (b, m @ GrpcMessageSchema(name, _, _, nested, nestedEnums)) =>
+        (b :+ (p.resolve(name) -> m)) ++
+          makeDictTypes(p :+ name, nested.getOrElse(Nil)) ++
+          makeDictTypes(p :+ name, nestedEnums.getOrElse(Nil))
+      case (b, m @ GrpcEnumSchema(name, _)) =>
+        b :+ (p.resolve(name) -> m)
+    }
 
   def validate(methodDescription: GrpcMethodDescription)(
       requestClass: String,
